Restructure the application,
Fixes error while running the front end and back end server in a production environment, Makes it easier to run the express server and create react app in development side by side
