--- backend/router/listing.js
@@ -1,6 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-router.get('/listing', (req, res, next)=> res.status(200).json({message: 'you have reached the listing page'}));
-
-module.exports = router;
