+++ package.json
@@ -10,7 +10,9 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.1"
+    "react-scripts": "3.4.1",
+    "express": "^4.17.1",
+    "mongoose": "^5.9.12"
   },
   "scripts": {
     "start": "react-scripts start",
